Auth fixes

+++ src/main/java/com/api/API2/controller/ReplyController.java
@@ -1,16 +1,14 @@
 package com.api.API2.controller;
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.*;
 import com.api.API2.entity.Reply;
 import com.api.API2.entity.User;
 import com.api.API2.service.ReplyService;
 import com.api.API2.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.EntityModel;
-import org.springframework.hateoas.Link;
 import org.springframework.hateoas.server.mvc.WebMvcLinkBuilder;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
@@ -31,6 +29,11 @@ public class ReplyController {
     @GetMapping("/Replies")
         public ResponseEntity<List<Reply>> getReplies(HttpServletRequest request){
 
+        String authorizationHeader = request.getHeader("Authorization");
+        if (authorizationHeader == null || authorizationHeader.isBlank()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         User u=userService.getUserById(request.getHeader("Authorization").substring(7));
         if(u!=null) {
 
@@ -39,18 +42,32 @@ public class ReplyController {
         return ResponseEntity.notFound().build();
     }
     @PostMapping("/reply/{th}")
-    public  ResponseEntity<Reply> postReply(@Valid @RequestBody Reply r, HttpServletRequest request, @PathVariable String th){
+    public ResponseEntity<EntityModel<Reply>> postReply(@Valid @RequestBody Reply r, HttpServletRequest request, @PathVariable String th){
+
+        String authorizationHeader = request.getHeader("Authorization");
+        if (authorizationHeader == null || authorizationHeader.isBlank()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         User u=userService.getUserById(request.getHeader("Authorization").substring(7));
         if(u!=null) {
             r.setUser(u);
             Reply savedReply = replyService.saveReply(r, th);
-            return ResponseEntity.created(URI.create("/reply/" + savedReply.getId())).body(savedReply);
+            String jwt = request.getHeader("Authorization").substring(7);
+            EntityModel<Reply> resource = EntityModel.of(r);
+            String concatThJwt=jwt+":"+savedReply.getThread().getId();
+            WebMvcLinkBuilder linkBuilder = WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(ThreadController.class).getAllThreadsp(concatThJwt, request));
+            resource.add(linkBuilder.withRel("thread"));
+            return ResponseEntity.created(URI.create("/reply/" + savedReply.getId())).body(resource);
         }
         return ResponseEntity.notFound().build();
     }
 
     @GetMapping("/reply/{id}")
     public ResponseEntity<EntityModel<Optional<Reply>>> getReplyById(@PathVariable String id, HttpServletRequest request) {
+        String authorizationHeader = request.getHeader("Authorization");
+        if (authorizationHeader == null || authorizationHeader.isBlank()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         Optional<Reply> r = replyService.getReplyById(id);
         String jwt = request.getHeader("Authorization").substring(7);
         EntityModel<Optional<Reply>> resource = EntityModel.of(r);
@@ -65,6 +82,12 @@ public class ReplyController {
 
     @DeleteMapping("/reply/{id}")
     public ResponseEntity<Object> deleteReply(@PathVariable String id, HttpServletRequest request) {
+
+        String authorizationHeader = request.getHeader("Authorization");
+        if (authorizationHeader == null || authorizationHeader.isBlank()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         User user = userService.getUserById(request.getHeader("Authorization").substring(7));
         if (user != null) {
             replyService.deleteReply(id, user.getId());
@@ -76,13 +99,22 @@ public class ReplyController {
 
 
     @PutMapping("/reply/{id}")
-    public ResponseEntity<Object> updateReply(@PathVariable String id,@RequestBody Reply r,HttpServletRequest request){
+    public ResponseEntity<Object> updateReply(@PathVariable String id, @RequestBody Reply r, HttpServletRequest request){
         Optional<Reply> replyOptional = replyService.getReplyById(id);
         if (replyOptional.isPresent()) {
             Reply existingReply = replyOptional.get();
             existingReply.setId(id); // Update the properties as needed
             existingReply.setAnswer(r.getAnswer());
-            return ResponseEntity.ok(replyService.saveReply(existingReply,existingReply.getThread().getId()));
+            String authorizationHeader = request.getHeader("Authorization");
+            if (authorizationHeader == null || authorizationHeader.isBlank()) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+            }
+            String jwt = request.getHeader("Authorization").substring(7);
+            EntityModel<Reply> resource = EntityModel.of(replyService.saveReply(existingReply,existingReply.getThread().getId()));
+            String concatThJwt=jwt+":"+resource.getContent().getThread().getId();
+            WebMvcLinkBuilder linkBuilder = WebMvcLinkBuilder.linkTo(WebMvcLinkBuilder.methodOn(ThreadController.class).getAllThreadsp(concatThJwt, request));
+            resource.add(linkBuilder.withRel("thread"));
+            return ResponseEntity.ok(resource);
         }
         return ResponseEntity.notFound().build();
     }
